+++ packages/ag-charts-community/src/util/baseModule.ts
@@ -1,0 +1,14 @@
+import type { JsonApplyParams } from './json';
+
+export interface ModuleInstance {
+    destroy(): void;
+}
+
+export interface BaseModule {
+    optionsKey: string;
+    packageType: 'community' | 'enterprise';
+    chartTypes: ('cartesian' | 'polar' | 'hierarchy')[];
+    identifier?: string;
+
+    optionConstructors?: JsonApplyParams['constructors'];
+}
